OSJ-110 and OSJ-111: Adjust handling in error cases.
In cases where configureBuilder() does throw, as it can now, make sure we leave things in
a known and acceptable state.
When injecting an external SchemaBuilder, clear the cached schema before configuring, that way
it's already cleared if the configure throws.
When in getSAMLSchema(), if lazy initing a new SchemaBuilder and the configure fails, clear
out the builder reference so it's not left in a half-inited state and used on the next call.